Time: 7 ms (8.92%) | Memory: 113.8 MB (26.28%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.cpp
@@ -3,11 +3,11 @@ public:
     double findMaxAverage(vector<int>& nums, int k) {
         int n = nums.size();
         if(k>n) return -1;
-        int sum = 0;
+        double sum = 0;
         for(int i = 0; i<k; i++){
             sum+=nums[i];
         }
-        int res = sum;
+        double res = sum;
         int low = 0;
         int high = k;
         while(high<n){
